LayoutTests/media/video-dom-loopend-expected.txt

Test that video loops when time reaches the value specified by the loopEnd DOM attribute.

EXPECTED (video.loopEnd == '0.5') OK
EXPECTED (video.getAttribute('loopend') == '0.5s') OK
RUN(video.play())
EVENT(timeupdate)
EVENT(timeupdate)
EXPECTED (video.currentLoop == '1') OK
EXPECTED (video.currentTime.toFixed(1) == '0') OK
END OF TEST
